Revert "feat(ui): extract design system into @velar/ui package"
