+++ apps/web/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import './globals.css';
 import { AppProviders } from '../components/AppProviders';
-import { themeInitScript } from '@velar/ui';
+import { themeInitScript } from '../components/ui/theme';
 
 export const metadata: Metadata = {
   title: 'VELAR | Trazabilidad de Bonos',
