+++ apps/web/components/AppProviders.tsx
@@ -2,22 +2,13 @@
 
 import type { ReactNode } from 'react';
 import { usePathname } from 'next/navigation';
-import { CountryProvider, useCountry } from '../lib/country';
+import { CountryProvider } from '../lib/country';
 import { WalletProvider } from '../lib/wallet';
-import { ThemeProvider } from '@velar/ui';
+import { ThemeProvider } from './ui/theme';
 import { ToastContainer } from './Toast';
 
 function isAuthRoute(pathname: string | null) {
   return pathname === '/login' || pathname === '/signup' || pathname === '/ir-login';
-}
-
-/**
- * Puente entre el contexto de país de la app y el design system: `@velar/ui` no
- * conoce `lib/country`, así que el país se le pasa por prop desde acá.
- */
-function ThemeWithCountry({ children }: { children: ReactNode }) {
-  const { country } = useCountry();
-  return <ThemeProvider country={country}>{children}</ThemeProvider>;
 }
 
 export function AppProviders({ children }: { children: ReactNode }) {
@@ -29,12 +20,12 @@ export function AppProviders({ children }: { children: ReactNode }) {
 
   return (
     <CountryProvider>
-      <ThemeWithCountry>
+      <ThemeProvider>
         <WalletProvider>
           {children}
           <ToastContainer />
         </WalletProvider>
-      </ThemeWithCountry>
+      </ThemeProvider>
     </CountryProvider>
   );
 }
