+++ apps/web/components/analytics/charts/StackedBarChart.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Bar, BarChart, CartesianGrid, Legend, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
-import { colorVar } from '@velar/ui';
+import { colorVar } from '../../ui/tokens';
 import { ChartFrame, ChartTableFallback, chartTooltipStyle } from './ChartFrame';
 
 export interface StackedSegmentDef {
