+++ apps/web/components/contract-engine/ContractEngineExplorer.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import type { AssembledContractDocument, ContractSummary, ContractVersionDiff, ContractVersionSummary } from '@velar/types';
-import { Alert, EmptyState, Spinner, Tabs } from '@velar/ui';
+import { Alert, EmptyState, Spinner, Tabs } from '../ui';
 import { createContractEngineClient, type FetchLike } from '../../lib/contract-engine';
 import { ContractSummaryPanel } from './ContractSummaryPanel';
 import { ContractDocumentPreview } from './ContractDocumentPreview';
