+++ apps/web/components/ui/Tooltip.tsx
@@ -4,7 +4,7 @@
  * CSS puro para posición; se muestra en focus para soporte de teclado.
  */
 import { useId, useState, type ReactNode } from 'react';
-import { cn } from './cn.js';
+import { cn } from './cn';
 
 type Side = 'top' | 'bottom' | 'left' | 'right';
