+++ apps/web/next.config.ts
@@ -3,7 +3,7 @@ import path from 'path';
 
 const nextConfig: NextConfig = {
   output: 'standalone',
-  transpilePackages: ['@velar/types', '@velar/ui'],
+  transpilePackages: ['@velar/types'],
   turbopack: {
     root: path.resolve(__dirname, '../..'),
   },
