--- packages/ui/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2017",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    /*
-     * ESM, no CommonJS: el emit CJS antepone "use strict" a la directiva
-     * 'use client' y Next deja de reconocer el boundary de cliente, que exige
-     * la directiva como primera sentencia del archivo.
-     */
-    "module": "esnext",
-    "moduleResolution": "bundler",
-    "jsx": "react-jsx",
-    "strict": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "declaration": true,
-    "outDir": "./dist"
-  },
-  "include": ["src"]
-}
